# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Customer service chat system API
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
        '500':
          description: Failed to generate upload URL.
...
  /agent/quick-replies:
    get:
      operationId: agentListQuickReplies
      tags: [agent]
      summary: List own quick replies
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search title or content
      responses:
        "200":
          description: Quick replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuickReply"
        "401":
          description: Not authenticated
    post:
      operationId: agentCreateQuickReply
      tags: [agent]
      summary: Create a quick reply
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuickReplyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickReply"
        "400":
          description: Validation error
        "401":
          description: Not authenticated
  /agent/quick-replies/{id}:
    patch:
      operationId: agentUpdateQuickReply
      tags: [agent]
      summary: Update own quick reply
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuickReplyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickReply"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: agentDeleteQuickReply
      tags: [agent]
      summary: Delete own quick reply
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/quick-replies:
    get:
      operationId: adminListQuickReplies
      tags: [admin]
      summary: List quick replies (all agents, super_admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Quick replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuickReplyWithAgent"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
    post:
      operationId: adminCreateQuickReply
      tags: [admin]
      summary: Create quick reply for an agent (super_admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateQuickReplyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickReplyWithAgent"
        "400":
          description: Validation error
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
  /admin/quick-replies/{id}:
    patch:
      operationId: adminUpdateQuickReply
      tags: [admin]
      summary: Update any quick reply (super_admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuickReplyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickReplyWithAgent"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: adminDeleteQuickReply
      tags: [admin]
      summary: Delete any quick reply (super_admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...
    QuickReply:
      type: object
      required: [id, agentId, title, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        agentId:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QuickReplyWithAgent:
      allOf:
        - $ref: "#/components/schemas/QuickReply"
        - type: object
          properties:
            agentDisplayName:
              type: ["string", "null"]
            agentUsername:
              type: ["string", "null"]
    CreateQuickReplyBody:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
    AdminCreateQuickReplyBody:
      type: object
      required: [agentId, title, content]
      properties:
        agentId:
          type: integer
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
    UpdateQuickReplyBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1